+++ src/main/java/kr/eolmago/service/deal/DealScheduler.java
@@ -1,4 +1,4 @@
-package kr.eolmago.scheduler;
+package kr.eolmago.service.deal;
 
 import kr.eolmago.domain.entity.deal.Deal;
 import kr.eolmago.domain.entity.deal.enums.DealStatus;
