+++ src/main/java/kr/eolmago/service/user/UserPenaltyScheduler.java
@@ -1,4 +1,4 @@
-package kr.eolmago.scheduler;
+package kr.eolmago.service.user;
 
 import kr.eolmago.domain.entity.user.User;
 import kr.eolmago.domain.entity.user.UserPenalty;
